requirements.txt
langchain==0.0.154
PyPDF2==3.0.1
python-dotenv==1.0.0
fastapi==0.68.1
uvicorn==0.15.0
httpx==0.19.0
starlette==0.15.0
faiss-cpu==1.7.4
altair<5
